Attempt to load syntax settings on startup
Instead of specifying the syntax settings in the plugin I want them
to be read from a .sublime-settings file. But, the sublime api I need
are not available when SublimeLinter loads the plugin. It does work
when reloading the plugin after sublime has opened, however. Trying to
see if I can use a dummy value and then force a reset using the
plugin_loaded function, which gets called after the plugin has loaded
and all sublime api's are available